+++ src/webhooks/java/gg/top/webhooks/entity/PartialProject.java
@@ -1,0 +1,30 @@
+package gg.top.webhooks.entity;
+
+import com.google.gson.annotations.SerializedName;
+
+public class PartialProject {
+  private String id;
+
+  private ProjectType type;
+
+  private Platform platform;
+
+  @SerializedName("platform_id")
+  private String platformId;
+
+  public String getId() {
+    return id;
+  }
+
+  public ProjectType getType() {
+    return type;
+  }
+
+  public Platform getPlatform() {
+    return platform;
+  }
+
+  public String getPlatformId() {
+    return platformId;
+  }
+}
