+++ src/webhooks/java/gg/top/webhooks/entity/Platform.java
@@ -1,0 +1,8 @@
+package gg.top.webhooks.entity;
+
+import com.google.gson.annotations.SerializedName;
+
+public enum Platform {
+  @SerializedName("discord")
+  DISCORD
+}
