+++ src/webhooks/java/gg/top/webhooks/entity/ProjectType.java
@@ -1,0 +1,11 @@
+package gg.top.webhooks.entity;
+
+import com.google.gson.annotations.SerializedName;
+
+public enum ProjectType {
+  @SerializedName("bot")
+  DISCORD_BOT,
+
+  @SerializedName("server")
+  DISCORD_SERVER
+}
