+++ src/webhooks/java/gg/top/webhooks/payload/VoteCreatePayload.java
@@ -1,0 +1,48 @@
+package gg.top.webhooks.payload;
+
+import java.time.OffsetDateTime;
+
+import com.google.gson.annotations.SerializedName;
+
+import gg.top.webhooks.entity.PartialProject;
+import gg.top.webhooks.entity.User;
+
+public class VoteCreatePayload {
+  private String id;
+
+  private int weight;
+
+  @SerializedName("created_at")
+  private OffsetDateTime votedAt;
+
+  @SerializedName("expires_at")
+  private OffsetDateTime expiresAt;
+
+  private PartialProject project;
+
+  private User user;
+
+  public String getId() {
+    return id;
+  }
+
+  public int getWeight() {
+    return weight;
+  }
+
+  public OffsetDateTime getVotedAt() {
+    return votedAt;
+  }
+
+  public OffsetDateTime getExpiredAt() {
+    return expiresAt;
+  }
+
+  public PartialProject getProject() {
+    return project;
+  }
+
+  public User getUser() {
+    return user;
+  }
+}
