--- src/webhooks/java/org/discordbots/webhooks/entity/Platform.java
@@ -1,8 +1,0 @@
-package org.discordbots.webhooks.entity;
-
-import com.google.gson.annotations.SerializedName;
-
-public enum Platform {
-  @SerializedName("discord")
-  DISCORD
-}
