--- src/webhooks/java/org/discordbots/webhooks/payload/IntegrationDeletePayload.java
@@ -1,12 +1,0 @@
-package org.discordbots.webhooks.payload;
-
-import com.google.gson.annotations.SerializedName;
-
-public class IntegrationDeletePayload {
-  @SerializedName("connection_id")
-  private String connectionId;
-
-  public String getConnectionId() {
-    return connectionId;
-  }
-}
